Hotfix: 헬스장 지역명 검증 제거및 프로필 이미지 검증 로직 수정

+++ api-server/src/main/java/com/lastone/apiserver/service/member/MemberServiceImpl.java
@@ -11,6 +11,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.ObjectUtils;
+import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
 import java.util.Optional;
@@ -34,7 +35,9 @@ public class MemberServiceImpl implements MemberService {
         isDuplicatedNickname(memberUpdateDto.getNickname(), member.getNickname());
 
         if (!ObjectUtils.isEmpty(profileImg)) {
-            s3Service.delete(member.getProfileUrl());
+            if (StringUtils.hasText(member.getProfileUrl())) {
+                s3Service.delete(member.getProfileUrl());
+            }
             memberUpdateDto.setProfileUrl(s3Service.upload(profileImg));
         }
         member.update(memberUpdateDto);
